Warning issues after fresh com_localise installation. (#47)
* Try to solve issue at my repo PRs

* Bad merge

* Warning issues after fresh installation

* Add JM code to solve the 'branches issue'

+++ administrator/components/com_localise/Field/BranchesField.php
@@ -53,14 +53,16 @@ class BranchesField extends ListField
 	{
 		$attributes    = '';
 		$params        = ComponentHelper::getParams('com_localise');
+		$branches      = array();
 		$branches_path = JPATH_ROOT
 				. '/administrator/components/com_localise/customisedref/joomla_branches.txt';
+		if (File::exists($branches_path))
+		{
+			file_put_contents($branches_path, '');
 
-		file_put_contents($branches_path, '');
-
-
-		$branches_file = file_get_contents($branches_path);
-		$branches      = preg_split("/\\r\\n|\\r|\\n/", $branches_file);
+			$branches_file = file_get_contents($branches_path);
+			$branches      = preg_split("/\\r\\n|\\r|\\n/", $branches_file);
+		}
 
 		$gh_user       = 'joomla';
 		$gh_project    = 'joomla-cms';
@@ -96,7 +98,7 @@ class BranchesField extends ListField
 			{
 				$branch_name = $active_branch->name;
 
-				if (!in_array($branch_name, $branches))
+				if (!in_array($branch_name, $branches) && is_numeric($branch_name[0]) && $branch_name[0] >= 4)
 				{
 					$branches[] = $branch_name;
 					Factory::getApplication()->enqueueMessage(
